Quite a bit of work since last commit,  - Added AppDisplay to show the sudoku grid as well as buttons for things like starting a new game or clearing the grid  - Added a CellGroup class for groups of (9) cells, which helps with making sure the rules are followed  - Added unit tests (mostly empty right now, but not all; need to work on them since they're good for catching small bugs)
The solver isn't very good right now; I think there's one or two more things I can add, and then might have to work on having it test values, that'll be tricky.

The GUI also needs a lot of work - it's not pretty at all, and it'd be nice to add other rendering features too.