I was going to add a third check that I would normally do, but writing the code for it started to get very involved so I left the class alone for now (CellGroup.MissingValue). Instead, it seemed easiest to add a recursive guessing solver - try a value in a cell, see if we can solve it then or if we get an error, in which case we go back and try a different potential value. Works with the test grid!
Also added a cell editor so the text looks decent when user is entering values and we also check that the value they enter is allowed, otherwise show an error. Those were the main things.

Next, I need to try to make the New Grid button actually create a valid sudoku grid...